+++ app/controllers/index.js
@@ -3,6 +3,7 @@ import Controller from '@ember/controller';
 import { computed } from '@ember/object';
 
 export default Controller.extend({
+
   currentSelector: computed(
     'esMujer',
     'esHombre',