+++ app/router.js
@@ -1,11 +1,28 @@
-import Router from '@ember/routing/router';
-// import config from './config/environment';
-// import googlePageview from './mixins/google-pageview';
+import EmberRouter from '@ember/routing/router';
+import RouterScroll from 'ember-router-scroll';
+import config from './config/environment';
+import { get } from '@ember/object';
+import { inject as service } from '@ember/service';
+import { scheduleOnce } from '@ember/runloop';
 
-// const Router = EmberRouter.extend(googlePageview, {
-//   location: config.locationType,
-//   rootURL: config.rootURL
-// });
+const Router = EmberRouter.extend(RouterScroll, {
+  location: config.locationType,
+  metrics: service(),
+
+  didTransition() {
+    this._super(...arguments);
+    this._trackPage();
+  },
+
+  _trackPage() {
+    scheduleOnce('afterRender', this, () => {
+      const page = this.get('url');
+      const title = this.getWithDefault('currentRouteName', 'unknown');
+
+      get(this, 'metrics').trackPage({ page, title });
+    });
+  }
+});
 
 Router.map(function() {
   this.route('institucion', function() {