+++ app/routes/index.js
@@ -54,45 +54,5 @@ export default Route.extend({
         });
       }
     });
-  },
-
-  /**
-   * Acciones de la ruta: applyFilter, applyShuffle. Pendiente pasar a controlador.
-   *
-   * @property actions
-   * @type {Object}
-   */
-  actions: {
-
-    // TODO: Pendiente de re-habilitar: esta función aplica un selector para el
-    // filtro de funcionarios
-    applyFilter(selector) {
-
-      var $container = $('#portfolio');
-
-      $('#portfolio-filter li').removeClass('activeFilter');
-
-      $('#' + selector).addClass('activeFilter');
-
-      var isotopeSelector = 'pf-todos' === selector ? '*' : '.' + selector;
-
-      $container.isotope({transitionDuration: '0.65s'});
-
-      $container.isotope({filter: isotopeSelector});
-
-      return false;
-    },
-
-    // TODO: Pendiente de re-habilitar: esta función aplica un shuffle a los items
-    // manejados por Isotope
-    applyShuffle() {
-      var $container = $('#portfolio');
-
-      $container.isotope({transitionDuration: '0.65s'});
-
-      $container.isotope('updateSortData').isotope({
-        sortBy: 'random'
-      });
-    }
   }
 });